# docs/documentacion/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgroFlow V2 - Dev",
    "description": "Reinicio de AgroFlow con estándares profesionales. Rama DEV.",
    "version": "2.0.0"
  },
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login",
        "description": "Login con Diagnóstico de Errores.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Me",
        "operationId": "me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/cambiar-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Cambiar Password",
        "description": "Permite al usuario cambiar su propia contraseña.",
        "operationId": "cambiar_password_api_v1_auth_cambiar_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CambiarPasswordPropiaBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/usuarios": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Listar Usuarios",
        "description": "Lista todos los usuarios (Solo Admin).",
        "operationId": "listar_usuarios_api_v1_auth_usuarios_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UsuarioResponse"
                  },
                  "type": "array",
                  "title": "Response Listar Usuarios Api V1 Auth Usuarios Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Crear Usuario",
        "description": "Crea un nuevo usuario (Solo Admin).",
        "operationId": "crear_usuario_api_v1_auth_usuarios_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/usuarios/{id}": {
      "patch": {
        "tags": [
          "Auth"
        ],
        "summary": "Actualizar Usuario",
        "description": "Actualiza datos de un usuario (Solo Admin).",
        "operationId": "actualizar_usuario_api_v1_auth_usuarios__id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/usuarios/{id}/reset-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Reset Password",
        "description": "Resetea la contraseña de un usuario a 123456 (Solo Admin).",
        "operationId": "reset_password_api_v1_auth_usuarios__id__reset_password_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestablecerPasswordBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/usuarios/{id}/desbloquear": {
      "patch": {
        "tags": [
          "Auth"
        ],
        "summary": "Desbloquear Usuario",
        "description": "Desbloquea a un usuario (Solo Admin).",
        "operationId": "desbloquear_usuario_api_v1_auth_usuarios__id__desbloquear_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/roles": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Listar Roles",
        "description": "Lista todos los roles maestros.",
        "operationId": "listar_roles_api_v1_auth_roles_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RolResponse"
                  },
                  "type": "array",
                  "title": "Response Listar Roles Api V1 Auth Roles Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Crear Rol",
        "description": "Crea un nuevo rol maestro.",
        "operationId": "crear_rol_api_v1_auth_roles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RolCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RolResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/roles/{id}": {
      "patch": {
        "tags": [
          "Auth"
        ],
        "summary": "Actualizar Rol",
        "description": "Actualiza un rol maestro.",
        "operationId": "actualizar_rol_api_v1_auth_roles__id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RolUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RolResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CambiarPasswordPropiaBody": {
        "properties": {
          "password_actual": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password Actual"
          },
          "nueva_password": {
            "type": "string",
            "title": "Nueva Password"
          }
        },
        "type": "object",
        "required": [
          "nueva_password"
        ],
        "title": "CambiarPasswordPropiaBody",
        "description": "Body del endpoint POST /api/v1/auth/cambiar-password-propia.\nPermite al usuario autenticado cambiar su propia contraseña.\n\nLógica de 'password_actual':\n  - Si 'requiere_cambio_password=True' (primer login o reset reciente),\n    NO se exige 'password_actual' porque el usuario no tiene una real.\n  - En cualquier otro caso, sí se exige para confirmar que quien cambia\n    la contraseña es efectivamente el dueño de la cuenta."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "usuario": {
            "type": "string",
            "title": "Usuario"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "usuario",
          "password"
        ],
        "title": "LoginRequest",
        "description": "Datos que el cliente envía al endpoint POST /api/v1/auth/login.\nAmbos campos son obligatorios. El router los valida con Pydantic\nantes de ejecutar cualquier lógica de negocio."
      },
      "LoginResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "usuario": {
            "type": "string",
            "title": "Usuario"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "rol": {
            "type": "string",
            "title": "Rol"
          },
          "requiere_cambio_password": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requiere Cambio Password",
            "default": false
          },
          "permisos": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Permisos"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "usuario",
          "nombre",
          "rol"
        ],
        "title": "LoginResponse",
        "description": "Respuesta exitosa del endpoint de login.\nContiene el token JWT y los datos básicos del usuario para que\nel frontend pueda mostrar el nombre y ajustar la UI según el rol,\nsin necesidad de hacer una segunda llamada a /me.\n\nNota sobre 'requiere_cambio_password':\n  Si es True, el frontend debe redirigir al usuario a la pantalla\n  de cambio de contraseña antes de permitirle acceder al sistema."
      },
...
      "MeResponse": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "usuario": {
            "type": "string",
            "title": "Usuario"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "rol": {
            "type": "string",
            "title": "Rol"
          },
          "requiere_cambio_password": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requiere Cambio Password",
            "default": false
          },
          "permisos": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Permisos"
          }
        },
        "type": "object",
        "required": [
          "usuario",
          "nombre",
          "rol"
        ],
        "title": "MeResponse",
        "description": "Respuesta del endpoint GET /api/v1/auth/me.\nRetorna los datos del usuario autenticado según el token JWT.\nEste endpoint es el que el frontend debe usar para obtener el rol\ndel usuario — NUNCA leer el rol desde variables de entorno del cliente.\n\n'from_attributes = True' permite convertir directamente el modelo\nORM (SQLAlchemy) a este schema Pydantic sin mapeo manual."
      },
...
      "RestablecerPasswordBody": {
        "properties": {
          "nueva_password": {
            "type": "string",
            "title": "Nueva Password",
            "default": "123456"
          }
        },
        "type": "object",
        "title": "RestablecerPasswordBody",
        "description": "Body del endpoint PATCH /api/v1/auth/usuarios/{id}/password-reset.\nSolo el administrador puede usar este endpoint para resetear la\ncontraseña de cualquier usuario.\n\nTras el reset, el sistema marca 'requiere_cambio_password=True' en el\nusuario, forzándolo a elegir una nueva contraseña en su próximo login."
      },
      "RolCreate": {
        "properties": {
          "nombre_rol": {
            "type": "string",
            "title": "Nombre Rol"
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion"
          },
          "permisos_plantilla": {
            "additionalProperties": true,
            "type": "object",
            "title": "Permisos Plantilla",
            "default": {
              "logicapture": true,
              "maestros": true,
              "operaciones": true,
              "sistema": false
            }
          }
        },
        "type": "object",
        "required": [
          "nombre_rol"
        ],
        "title": "RolCreate"
      },
      "RolResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "nombre_rol": {
            "type": "string",
            "title": "Nombre Rol"
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion"
          },
          "permisos_plantilla": {
            "additionalProperties": true,
            "type": "object",
            "title": "Permisos Plantilla"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nombre_rol",
          "permisos_plantilla"
        ],
        "title": "RolResponse"
      },
      "RolUpdate": {
        "properties": {
          "nombre_rol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nombre Rol"
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion"
          },
          "permisos_plantilla": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Permisos Plantilla"
          }
        },
        "type": "object",
        "title": "RolUpdate"
      },
...
      "UsuarioCreate": {
        "properties": {
          "usuario": {
            "type": "string",
            "title": "Usuario"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "password": {
            "type": "string",
            "title": "Password",
            "default": "123456"
          },
          "rol": {
            "type": "string",
            "title": "Rol"
          },
          "permisos": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Permisos",
            "default": {
              "logicapture": true,
              "maestros": true,
              "operaciones": true,
              "sistema": false
            }
          }
        },
        "type": "object",
        "required": [
          "usuario",
          "nombre",
          "rol"
        ],
        "title": "UsuarioCreate",
        "description": "Datos para crear un nuevo usuario. Solo el administrador puede usar\nel endpoint que recibe este schema (POST /api/v1/auth/usuarios).\n\n'password' tiene valor por defecto '123456' porque tras la creación\nel sistema asigna 'requiere_cambio_password=True', obligando al nuevo\nusuario a cambiar su contraseña en el primer inicio de sesión."
      },
      "UsuarioResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "usuario": {
            "type": "string",
            "title": "Usuario"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "rol": {
            "type": "string",
            "title": "Rol"
          },
          "activo": {
            "type": "boolean",
            "title": "Activo"
          },
          "requiere_cambio_password": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requiere Cambio Password",
            "default": false
          },
          "permisos": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Permisos"
          },
          "bloqueado": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bloqueado",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "usuario",
          "nombre",
          "rol",
          "activo"
        ],
        "title": "UsuarioResponse",
        "description": "Representación pública de un usuario. Nunca incluye password_hash\nni otros campos sensibles. Usada en todas las respuestas que\ndevuelven datos de usuario al cliente."
      },
      "UsuarioUpdate": {
        "properties": {
          "nombre": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nombre"
          },
          "rol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rol"
          },
          "usuario": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usuario"
          },
          "permisos": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Permisos"
          },
          "activo": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activo"
          }
        },
        "type": "object",
        "title": "UsuarioUpdate",
        "description": "Datos para actualizar un usuario existente (PATCH /api/v1/auth/usuarios/{id}).\nTodos los campos son opcionales: solo se actualizan los que se envíen.\nEsto implementa el patrón \"partial update\" — más seguro que reemplazar\ntodo el objeto porque evita borrar campos no enviados por accidente."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}